chore: Added universal_io package for web support
